+++ lib/user.js
@@ -11,7 +11,6 @@ const Mailer = require('pixl-mail');
 const Request = require('pixl-request');
 const bcrypt = require('bcryptjs');
 const { Client } = require('ldapts');
-const OIDC = require('./oidc');
 
 const RETURN_USER_OBJECT = true
 const VERIFY_GROUP_MEMBERSHIP = true
@@ -20,6 +19,9 @@ module.exports = Class.create({
 
 	__name: 'User',
 	__parent: Component,
+	__mixins: [
+		require('./oidc') // OpenID Connect Layer Mixin
+	],
 
 	defaultConfig: {
 		"smtp_hostname": "",
@@ -90,18 +92,7 @@ module.exports = Class.create({
 		this.storage = this.server.Storage;
 		this.web = this.server.WebServer;
 
-		// OIDC sessions have related persistent index records which must be
-		// removed when the storage maintenance job expires the session.
-		this.storage.addRecordType('oidc_session', {
-			'delete': this.deleteExpiredOidcSession.bind(this)
-		});
-
-		this.installOidcEarlyRedaction();
-
-		// pixl-server-api may log parsed parameters in debug mode.  Move secrets
-		// to non-enumerable request properties before the API logger sees them.
-		this.web.addURIFilter(/\/api\/user\/(?:oauth|callback|logout|oidc_logout_redirect|oidc_backchannel_logout)(?:\?|$)/,
-			'OIDC Secret Filter', this.filterOidcRequest.bind(this));
+		this.setupOidc();
 
 		// hook system for integrating with outer webapp
 		this.hooks = {};
@@ -112,140 +103,10 @@ module.exports = Class.create({
 		this.usernameBlock = new RegExp(this.config.get('block_username_match'), "i");
 
 		// ------------------------------------  OAUTH ------------------------------------------
-		this.oauth_state = {}  // simple in-memory store for short-lived "state" parameter. 
-		this.oidc_logout_ticket_secret = Tools.generateUniqueID(64);
-		this.oidc_logout_ticket_replay = new Set();
-		this.server.on('day', ()=>{
-			self.oauth_state = {}
-			self.oidc_logout_ticket_replay.clear();
-		})  // clear stale oatuh request on daily basis	
+		this.oauth_state = {}  // simple in-memory store for short-lived "state" parameter.
+		this.server.on('day', function() { self.oauth_state = {}; });
 		// startup complete
 		callback();
-	},
-
-	installOidcEarlyRedaction: function() {
-		const web = this.web;
-		if (web._oidcEarlyRedactionInstalled) return;
-		web._oidcEarlyRedactionInstalled = true;
-		const enqueue = web.enqueueHTTPRequest;
-		const sensitivePath = /\/api\/user\/(?:oauth|callback|logout|oidc_logout_redirect|oidc_backchannel_logout)$/;
-
-		web.enqueueHTTPRequest = function(request, response) {
-			let parsed;
-			try { parsed = new URL(request.url, 'http://localhost'); }
-			catch (err) { return enqueue.call(web, request, response); }
-			if (!sensitivePath.test(parsed.pathname)) return enqueue.call(web, request, response);
-
-			const secrets = {};
-			['code', 'state', 'session_id', 'ticket'].forEach(function(name) {
-				if (!parsed.searchParams.has(name)) return;
-				secrets[name] = parsed.searchParams.get(name);
-				parsed.searchParams.set(name, '[REDACTED]');
-			});
-			request.url = parsed.pathname + parsed.search + parsed.hash;
-
-			const headers = request.headers || {};
-			if (headers['x-session-id']) secrets.session_id = headers['x-session-id'];
-			if (headers.cookie) {
-				const match = String(headers.cookie).match(/(?:^|;\s*)session_id=([^;]+)/);
-				if (match && !secrets.session_id) {
-					try { secrets.session_id = decodeURIComponent(match[1]); }
-					catch (err) { secrets.session_id = match[1]; }
-				}
-			}
-			['authorization', 'cookie', 'cf-access-jwt-assertion', 'x-api-key', 'x-session-id'].forEach(function(name) {
-				if (headers[name]) headers[name] = '[REDACTED]';
-			});
-			Object.defineProperty(request, '_oidc_early_secrets', {
-				value: secrets,
-				enumerable: false,
-				configurable: false
-			});
-			return enqueue.call(web, request, response);
-		};
-	},
-
-	filterOidcRequest: function(args, callback) {
-		const path = String(args.request.url || '').replace(/\?.*$/, '');
-		const action = path.split('/').pop();
-		const params = args.params || {};
-		const query = args.query || {};
-		const headers = args.request.headers || {};
-		const cookies = args.cookies || {};
-		const early = args.request._oidc_early_secrets || {};
-		const secrets = {};
-
-		if (action === 'callback') {
-			secrets.code = early.code || params.code || query.code;
-			secrets.state = early.state || params.state || query.state;
-			if (params.code) params.code = '[REDACTED]';
-			if (params.state) params.state = '[REDACTED]';
-			if (query.code) query.code = '[REDACTED]';
-			if (query.state) query.state = '[REDACTED]';
-		}
-
-		if (action === 'oidc_backchannel_logout') {
-			secrets.logout_token = params.logout_token;
-			if (Object.prototype.hasOwnProperty.call(params, 'logout_token')) params.logout_token = '[REDACTED]';
-		}
-
-		if (action === 'oidc_logout_redirect') {
-			secrets.ticket = early.ticket || params.ticket || query.ticket;
-			if (params.ticket) params.ticket = '[REDACTED]';
-			if (query.ticket) query.ticket = '[REDACTED]';
-		}
-
-		if (action === 'oauth' || action === 'logout') {
-			secrets.session_id = early.session_id || cookies.session_id || headers['x-session-id'] || params.session_id || query.session_id;
-			if (cookies.session_id) cookies.session_id = '[REDACTED]';
-			if (headers['x-session-id']) headers['x-session-id'] = '[REDACTED]';
-			if (params.session_id) params.session_id = '[REDACTED]';
-			if (query.session_id) query.session_id = '[REDACTED]';
-		}
-
-		['authorization', 'cookie', 'cf-access-jwt-assertion', 'x-api-key'].forEach((name) => {
-			if (headers[name]) headers[name] = '[REDACTED]';
-		});
-		Object.keys(cookies).forEach((name) => { cookies[name] = '[REDACTED]'; });
-
-		Object.defineProperty(args, '_oidc_secrets', {
-			value: secrets,
-			enumerable: false,
-			configurable: false
-		});
-
-		if (action === 'oidc_logout_redirect') {
-			if (args.request.method !== 'GET') {
-				return callback('405 Method Not Allowed', { 'Content-Type': 'text/plain', 'Allow': 'GET', 'Cache-Control': 'no-store' },
-					'Method not allowed.');
-			}
-			let payload;
-			try {
-				payload = OIDC.consumeLogoutTicket(secrets.ticket, this.getLogoutTicketSecret());
-				if (this.oidc_logout_ticket_replay.has(payload.jti)) throw new Error('Logout ticket was already used');
-				OIDC.parseSecureUrl(payload.location, 'OIDC logout redirect', true);
-				this.oidc_logout_ticket_replay.add(payload.jti);
-			}
-			catch (err) {
-				this.logError('oauth', 'Invalid OIDC logout redirect ticket');
-				return callback('400 Bad Request', { 'Content-Type': 'text/plain', 'Cache-Control': 'no-store' },
-					'Invalid logout request.');
-			}
-
-			args.response.writeHead(302, {
-				'Location': payload.location,
-				'Cache-Control': 'no-store',
-				'Referrer-Policy': 'no-referrer'
-			});
-			args.response.end();
-			return callback(true);
-		}
-		callback(false);
-	},
-
-	getLogoutTicketSecret: function() {
-		return (this.server && this.server.config && this.server.config.get('secret_key')) ||
-			this.oidc_logout_ticket_secret;
 	},
 
 
@@ -681,170 +542,8 @@ module.exports = Class.create({
 			args.params.session_id || args.query.session_id;
 	},
 
-	getOidcIndexEntries: function(session) {
-		if (!session || session.auth_provider !== 'oidc' || !session.oidc_issuer) return [];
-		const entries = [];
-		if (session.oidc_sid) entries.push({ type: 'sid', value: session.oidc_sid });
-		if (session.oidc_subject) entries.push({ type: 'sub', value: session.oidc_subject });
-		return entries;
-	},
-
-	withOidcRevocationLocks: function(keys, worker, callback) {
-		const self = this;
-		keys = Array.from(new Set(keys)).sort();
-		const acquired = [];
-		let pos = 0;
-		let released = false;
-
-		const release = function(err, value) {
-			if (released) return;
-			released = true;
-			while (acquired.length) self.storage.unlock('oidc-revocation-lock/' + acquired.pop());
-			callback(err || null, value);
-		};
-		const acquire = function() {
-			if (pos >= keys.length) {
-				try { return worker(release); }
-				catch (err) { return release(err); }
-			}
-			const key = keys[pos++];
-			self.storage.lock('oidc-revocation-lock/' + key, true, function() {
-				acquired.push(key);
-				acquire();
-			});
-		};
-		acquire();
-	},
-
-	findActiveOidcRevocation: function(keys, callback) {
-		const self = this;
-		const now = Tools.timeNow(true);
-		let pos = 0;
-		const next = function(err) {
-			if (err) return callback(err);
-			if (pos >= keys.length) return callback(null, null);
-			const key = keys[pos++];
-			self.storage.get(key, function(getErr, marker) {
-				if (getErr && self.isMissingStorageError(getErr)) return next();
-				if (getErr) return callback(getErr);
-				if (!marker || !Number.isFinite(marker.expires)) {
-					return callback(new Error('Invalid OIDC logout revocation marker'));
-				}
-				if (marker.expires >= now) return callback(null, marker);
-				self.storage.delete(key, function(deleteErr) {
-					if (self.isMissingStorageError(deleteErr)) deleteErr = null;
-					next(deleteErr);
-				});
-			});
-		};
-		next();
-	},
-
-	mutateOidcIndexKey: function(key, sessionId, expires, add, callback) {
-		const self = this;
-		const lockKey = 'oidc-index-lock/' + key;
-
-		self.storage.lock(lockKey, true, function() {
-			let finished = false;
-			const finish = function(err) {
-				if (finished) return;
-				finished = true;
-				self.storage.unlock(lockKey);
-				callback(err || null);
-			};
-
-			self.storage.get(key, function(err, record) {
-				if (err && !self.isMissingStorageError(err)) return finish(err);
-				record = record || { sessions: {} };
-				if (!record.sessions || typeof record.sessions !== 'object' || Array.isArray(record.sessions)) {
-					return finish(new Error('Invalid OIDC session index record'));
-				}
-
-				if (add) record.sessions[sessionId] = expires || 0;
-				else delete record.sessions[sessionId];
-
-				if (!Object.keys(record.sessions).length) {
-					if (err) return finish();
-					return self.storage.delete(key, function(deleteErr) {
-						if (self.isMissingStorageError(deleteErr)) deleteErr = null;
-						finish(deleteErr);
-					});
-				}
-
-				self.storage.put(key, record, finish);
-			});
-		});
-	},
-
-	updateOidcSessionIndexes: function(session, add, callback) {
-		const self = this;
-		const entries = self.getOidcIndexEntries(session);
-		let pos = 0;
-		const completed = [];
-
-		const next = function(err) {
-			if (err) {
-				if (!completed.length) return callback(err);
-				let rollbackPos = completed.length - 1;
-				let rollbackError = null;
-				const rollback = function(currentError) {
-					if (currentError && !rollbackError) rollbackError = currentError;
-					if (rollbackPos < 0) {
-						if (rollbackError) return callback(new Error(err.message + '; index rollback failed: ' + rollbackError.message));
-						return callback(err);
-					}
-					const entry = completed[rollbackPos--];
-					self.mutateOidcIndexKey(
-						OIDC.indexKey(session.oidc_issuer, entry.type, entry.value),
-						session.id, session.expires, !add, rollback
-					);
-				};
-				return rollback();
-			}
-			if (pos >= entries.length) return callback();
-			const entry = entries[pos++];
-			self.mutateOidcIndexKey(
-				OIDC.indexKey(session.oidc_issuer, entry.type, entry.value),
-				session.id, session.expires, add,
-				function(indexErr) {
-					if (!indexErr) completed.push(entry);
-					next(indexErr);
-				}
-			);
-		};
-		next();
-	},
-
-	storeOidcSession: function(session, callback) {
-		const self = this;
-		session.type = 'oidc_session';
-		const revocationKeys = self.getOidcIndexEntries(session).map(function(entry) {
-			return OIDC.revocationKey(session.oidc_issuer, session.oidc_client_id, entry.type, entry.value);
-		});
-		if (!session.oidc_client_id || !revocationKeys.length) {
-			return callback(new Error('OIDC session is missing revocation identity metadata'));
-		}
-		self.withOidcRevocationLocks(revocationKeys, function(done) {
-			self.findActiveOidcRevocation(revocationKeys, function(revocationErr, marker) {
-				if (revocationErr) return done(revocationErr);
-				if (marker) return done(new Error('OIDC session was revoked before it could be stored'));
-				self.storage.put('sessions/' + session.id, session, function(err) {
-					if (err) return done(err);
-					self.updateOidcSessionIndexes(session, true, function(indexErr) {
-						if (!indexErr) {
-							self.storage.expire('sessions/' + session.id, session.expires);
-							return done();
-						}
-						self.storage.delete('sessions/' + session.id, function(deleteErr) {
-							done(indexErr || deleteErr);
-						});
-					});
-				});
-			});
-		}, callback);
-	},
-
 	deleteSessionRecord: function(session, callback) {
+		// Delete OIDC indexes together with their primary session record.
 		const self = this;
 		if (!session || !session.id) return callback();
 		if (session.auth_provider !== 'oidc') {
@@ -882,78 +581,6 @@ module.exports = Class.create({
 		});
 	},
 
-	deleteExpiredOidcSession: function(key, session, callback) {
-		// Called by pixl-server-storage maintenance for an expired custom record.
-		this.deleteSessionRecord(session, callback);
-	},
-
-	processLogoutJti: function(issuer, jti, expires, worker, callback) {
-		const self = this;
-		const key = OIDC.replayKey(issuer, jti);
-		const lockKey = 'oidc-replay-lock/' + key;
-		self.storage.lock(lockKey, true, function() {
-			const finish = function(err, replay) {
-				self.storage.unlock(lockKey);
-				callback(err || null, !!replay);
-			};
-			self.storage.get(key, function(err) {
-				if (!err) return finish(null, true);
-				if (!self.isMissingStorageError(err)) return finish(err);
-				worker(function(workerErr) {
-					if (workerErr) return finish(workerErr);
-					self.storage.put(key, { created: Tools.timeNow(true), expires: expires }, function(putErr) {
-						if (putErr) return finish(putErr);
-						self.storage.expire(key, expires);
-						finish(null, false);
-					});
-				});
-			});
-		});
-	},
-
-	revokeOidcSessionsForClaim: function(issuer, claimType, claimValue, clientId, expires, callback) {
-		const self = this;
-		const key = OIDC.indexKey(issuer, claimType, claimValue);
-		const revocationKey = OIDC.revocationKey(issuer, clientId, claimType, claimValue);
-		self.withOidcRevocationLocks([revocationKey], function(done) {
-			self.storage.put(revocationKey, { created: Tools.timeNow(true), expires: expires }, function(markerErr) {
-				if (markerErr) return done(markerErr);
-				self.storage.expire(revocationKey, expires);
-				self.storage.get(key, function(err, record) {
-					if (err && self.isMissingStorageError(err)) return done(null, 0);
-					if (err) return done(err);
-					if (!record || !record.sessions || typeof record.sessions !== 'object') {
-						return done(new Error('Invalid OIDC session index record'));
-					}
-
-					const ids = Object.keys(record.sessions);
-					let pos = 0;
-					let deleted = 0;
-					const next = function(deleteErr) {
-						if (deleteErr) return done(deleteErr);
-						if (pos >= ids.length) return done(null, deleted);
-						const id = ids[pos++];
-						self.storage.get('sessions/' + id, function(sessionErr, session) {
-							if (sessionErr && self.isMissingStorageError(sessionErr)) {
-								return self.mutateOidcIndexKey(key, id, 0, false, next);
-							}
-							if (sessionErr) return next(sessionErr);
-							const identityMatches = session.auth_provider === 'oidc' && session.oidc_issuer === issuer &&
-								((claimType === 'sid' && session.oidc_sid === claimValue) ||
-								(claimType === 'sub' && session.oidc_subject === claimValue));
-							if (!identityMatches) return self.mutateOidcIndexKey(key, id, 0, false, next);
-							if (session.oidc_client_id !== clientId) return next();
-							self.deleteSessionRecord(session, function(removeErr) {
-								if (!removeErr) deleted++;
-								next(removeErr);
-							});
-						});
-					};
-					next();
-				});
-			});
-		}, callback);
-	},
 
 	// helper to join base_path with given location
 	getBaseLocation: function (subPath) {
@@ -1123,7 +750,7 @@ module.exports = Class.create({
 				throw new Error('OIDC token endpoint did not return an ID Token required for RP-initiated logout');
 			}
 			if (tokenResponse.id_token && oidcConfigured) {
-				const verified = await OIDC.verifyIdToken(tokenResponse.id_token, oauth, stateData.nonce);
+				const verified = await self.verifyOidcIdToken(tokenResponse.id_token, oauth, stateData.nonce);
 				oidcClaims = verified.payload;
 			}
 
@@ -1151,12 +778,12 @@ module.exports = Class.create({
 				throw new Error('OIDC UserInfo subject does not match the ID Token subject');
 			}
 			if (!oidcClaims && oauth.issuer && oauth.jwks_url && userResponse.sub) {
-				const provider = OIDC.validateProviderConfig(oauth);
+				const provider = self.validateOidcProviderConfig(oauth);
 				oidcClaims = { iss: provider.issuer, sub: userResponse.sub };
 			}
 			if (oauth.issuer && oauth.jwks_url) {
-				oidcMetadata = OIDC.buildSessionMetadata(oauth, oidcClaims, userResponse, tokenResponse.id_token,
-					self.getLogoutTicketSecret());
+				oidcMetadata = self.buildOidcSessionMetadata(oauth, oidcClaims, userResponse, tokenResponse.id_token,
+					self.getOidcLogoutTicketSecret());
 			}
 
 			// ==== Check if user exist in user database
@@ -1443,7 +1070,7 @@ module.exports = Class.create({
 					if (session.auth_provider === 'oidc') {
 						const oauth = self.getOauthConfig();
 						if (!oauth) throw new Error('OAuth configuration is unavailable');
-						logoutLocation = OIDC.buildLogoutLocation(oauth, session, self.getLogoutTicketSecret());
+						logoutLocation = self.buildOidcLogoutLocation(oauth, session, self.getOidcLogoutTicketSecret());
 					}
 				}
 				catch (configErr) {
@@ -1463,7 +1090,7 @@ module.exports = Class.create({
 					const response = { code: 0 };
 					if (logoutLocation) {
 						try {
-							const ticket = OIDC.createLogoutTicket(logoutLocation, self.getLogoutTicketSecret(), 60);
+							const ticket = self.createOidcLogoutTicket(logoutLocation, self.getOidcLogoutTicketSecret(), 60);
 							response.logout_location = self.getBaseLocation('/api/user/oidc_logout_redirect?ticket=' +
 								encodeURIComponent(ticket));
 						}
@@ -1481,68 +1108,6 @@ module.exports = Class.create({
 		}); // load session
 	},
 
-	api_oidc_backchannel_logout: async function(args, callback) {
-		const self = this;
-		const badRequest = function() {
-			callback('400 Bad Request', { 'Content-Type': 'application/json', 'Cache-Control': 'no-store' },
-				JSON.stringify({ code: 1, description: 'Invalid logout request.' }));
-		};
-		const serverError = function(status, description) {
-			callback(status, { 'Content-Type': 'application/json', 'Cache-Control': 'no-store' },
-				JSON.stringify({ code: 1, description: description }));
-		};
-
-		if (args.request.method !== 'POST') {
-			return callback('405 Method Not Allowed', { 'Content-Type': 'application/json', 'Allow': 'POST', 'Cache-Control': 'no-store' },
-				JSON.stringify({ code: 1, description: 'Method not allowed.' }));
-		}
-		if (!String(args.request.headers['content-type'] || '').match(/^application\/x-www-form-urlencoded(?:\s*;|$)/i)) {
-			return badRequest();
-		}
-
-		const oauth = self.getOauthConfig();
-		if (!oauth || !(oauth.backchannel_logout || {}).enabled) return badRequest();
-		const maxTokenSize = Math.max(1024, Number(oauth.backchannel_logout.max_token_size) || 16384);
-		const contentLength = Number(args.request.headers['content-length']);
-		if (Number.isFinite(contentLength) && contentLength > maxTokenSize + 1024) return badRequest();
-
-		try { OIDC.validateProviderConfig(oauth); }
-		catch (configErr) {
-			self.logError('oauth', 'OIDC back-channel logout is misconfigured: ' + configErr.message);
-			return serverError('500 Internal Server Error', 'OIDC logout is not configured correctly.');
-		}
-
-		const token = args._oidc_secrets && args._oidc_secrets.logout_token;
-		let verified;
-		try {
-			verified = await OIDC.verifyLogoutToken(token, oauth);
-		}
-		catch (verifyErr) {
-			const code = String(verifyErr.code || '');
-			const validationError = code.match(/^ERR_(?:JWT|JWS|JOSE|JWKS)/) ||
-				String(verifyErr.message || '').match(/(?:token|claim|signature|nonce|sid|subject|iat|events|size)/i);
-			self.logError('oauth', 'OIDC back-channel logout validation failed: ' + (code || verifyErr.message));
-			if (validationError) return badRequest();
-			return serverError('503 Service Unavailable', 'OIDC signing keys are temporarily unavailable.');
-		}
-
-		const payload = verified.payload;
-		const claimType = payload.sid ? 'sid' : 'sub';
-		const claimValue = payload.sid || payload.sub;
-
-		self.processLogoutJti(verified.provider.issuer, payload.jti, payload.exp, function(done) {
-			self.revokeOidcSessionsForClaim(verified.provider.issuer, claimType, claimValue,
-				verified.provider.clientId, payload.exp, done);
-		}, function(processErr, replay) {
-			if (processErr) {
-				self.logError('oauth', 'OIDC back-channel logout storage failure: ' + processErr.message);
-				return serverError('500 Internal Server Error', 'OIDC logout could not be completed.');
-			}
-			self.logDebug(6, replay ? 'OIDC back-channel logout replay handled idempotently' :
-				'OIDC back-channel logout completed');
-			callback({ code: 0 });
-		});
-	},
 
 	api_resume_session: function (args, callback) {
 		// validate existing session
@@ -1593,8 +1158,11 @@ module.exports = Class.create({
 
 				self.logDebug(6, "Recovering session for: " + session.username);
 
-				// store session object
-				self.storage.put('sessions/' + session.id, session, function (err, data) {
+				// OIDC resumes share the callback/logout locks so stale state cannot recreate a revoked session.
+				const storeSession = session.auth_provider === 'oidc' ?
+					self.storeOidcSession.bind(self) :
+					function(localSession, done) { self.storage.put('sessions/' + localSession.id, localSession, done); };
+				storeSession(session, function (err) {
 					if (err) {
 						return self.doError('user', "Failed to update session: " + err, callback);
 					}
