+++ test/oidc-logout.test.js
@@ -1,8 +1,8 @@
 const assert = require('node:assert/strict');
 const http = require('node:http');
 
-const OIDC = require('../lib/oidc');
 const User = require('../lib/user');
+const OIDC = User.prototype;
 const Admin = require('../lib/api/admin');
 
 let suiteBefore = function() {};
@@ -146,17 +146,28 @@ function invokeBackchannel(user, token, overrides) {
 		};
 		user.api_oidc_backchannel_logout(args, function(first, headers, body) {
 			if (typeof first === 'string') {
-				return resolve({ status: Number(first.split(' ')[0]), headers, body: JSON.parse(body) });
+				return resolve({ status: Number(first.split(' ')[0]), headers: headers || {}, body: body ? JSON.parse(body) : null });
 			}
-			resolve({ status: 200, body: first });
+			resolve({ status: 200, headers: {}, body: first });
 		});
+	});
+}
+
+function invokeLocalLogout(user, sessionId) {
+	return new Promise((resolve) => {
+		user.api_logout({
+			request: { headers: { 'user-agent': 'test' } },
+			response: { setHeader: function() {} },
+			params: {}, query: {}, cookies: {}, ip: '127.0.0.1',
+			_oidc_secrets: { session_id: sessionId }
+		}, resolve);
 	});
 }
 
 function recordExists(storage, key) { return storage.records.has(key); }
 
 test('RP logout is disabled by default', () => {
-	assert.equal(OIDC.buildLogoutLocation({ client_id: 'client' }, {}), null);
+	assert.equal(OIDC.buildOidcLogoutLocation({ client_id: 'client' }, {}), null);
 });
 
 test('RP logout builds a configuration-only URL without an ID Token hint', () => {
@@ -169,7 +180,7 @@ test('RP logout builds a configuration-only URL without an ID Token hint', () =>
 			params: { ui_locales: 'pl', post_logout_redirect_uri: 'https://attacker.example/' }
 		}
 	};
-	const url = new URL(OIDC.buildLogoutLocation(oauth, {}));
+	const url = new URL(OIDC.buildOidcLogoutLocation(oauth, {}));
 	assert.equal(url.origin, 'https://idp.example');
 	assert.equal(url.searchParams.get('post_logout_redirect_uri'), 'https://cron.example/');
 	assert.equal(url.searchParams.get('client_id'), 'cronicle-edge');
@@ -180,12 +191,22 @@ test('RP logout builds a configuration-only URL without an ID Token hint', () =>
 test('RP logout automatically includes a verified server-side ID Token hint', () => {
 	const oauth = {
 		client_id: 'cronicle-edge',
+		issuer: 'https://idp.example/',
+		jwks_url: 'https://idp.example/jwks',
 		logout: { enabled: true, end_session_url: 'https://idp.example/logout' }
 	};
-	const encrypted = OIDC.encryptSessionToken('signed.id.token', 'session-secret');
-	const url = new URL(OIDC.buildLogoutLocation(oauth, { oidc_id_token_hint_enc: encrypted }, 'session-secret'));
+	const encrypted = OIDC.encryptOidcSessionToken('signed.id.token', 'session-secret');
+	const session = {
+		auth_provider: 'oidc', oidc_issuer: oauth.issuer,
+		oidc_client_id: oauth.client_id, oidc_id_token_hint_enc: encrypted
+	};
+	const url = new URL(OIDC.buildOidcLogoutLocation(oauth, session, 'session-secret'));
 	assert.equal(url.searchParams.get('id_token_hint'), 'signed.id.token');
 	assert.equal(url.searchParams.get('client_id'), null);
+	assert.throws(() => OIDC.buildOidcLogoutLocation(Object.assign({}, oauth, { client_id: 'other-client' }),
+		session, 'session-secret'), /does not match/);
+	assert.throws(() => OIDC.buildOidcLogoutLocation(Object.assign({}, oauth, { issuer: 'https://other.example/' }),
+		session, 'session-secret'), /does not match/);
 });
 
 test('RP logout never silently falls back for an OIDC session without an ID Token hint', () => {
@@ -197,28 +218,64 @@ test('RP logout never silently falls back for an OIDC session without an ID Toke
 			post_logout_redirect_uri: 'https://cron.example/'
 		}
 	};
-	assert.throws(() => OIDC.buildLogoutLocation(oauth, { auth_provider: 'oidc' }, 'session-secret'),
+	assert.throws(() => OIDC.buildOidcLogoutLocation(oauth, { auth_provider: 'oidc' }, 'session-secret'),
 		/requires|does not contain a verified ID Token/);
 });
 
 test('Logout redirect tickets hide the ID Token and expire safely', () => {
 	const destination = 'https://idp.example/logout?id_token_hint=secret.id.token';
-	const ticket = OIDC.createLogoutTicket(destination, 'ticket-secret', 60);
+	const ticket = OIDC.createOidcLogoutTicket(destination, 'ticket-secret', 60);
 	assert.equal(ticket.includes('secret.id.token'), false);
-	assert.equal(OIDC.consumeLogoutTicket(ticket, 'ticket-secret').location, destination);
-	assert.throws(() => OIDC.consumeLogoutTicket(ticket, 'wrong-secret'));
+	assert.equal(OIDC.consumeOidcLogoutTicket(ticket, 'ticket-secret').location, destination);
+	assert.throws(() => OIDC.consumeOidcLogoutTicket(ticket, 'wrong-secret'));
 });
 
 test('RP logout rejects unsafe and request-controlled URLs', () => {
-	assert.throws(() => OIDC.buildLogoutLocation({ logout: {
+	assert.throws(() => OIDC.buildOidcLogoutLocation({ logout: {
 		enabled: true, end_session_url: 'javascript:alert(1)'
 	}}, {}), /HTTPS/);
-	assert.throws(() => OIDC.buildLogoutLocation({ logout: {
+	assert.throws(() => OIDC.buildOidcLogoutLocation({ logout: {
 		enabled: true, end_session_url: 'http://idp.example/logout', allow_http_localhost: true
 	}}, {}), /HTTPS/);
-	assert.doesNotThrow(() => OIDC.buildLogoutLocation({ logout: {
+	assert.doesNotThrow(() => OIDC.buildOidcLogoutLocation({ logout: {
 		enabled: true, end_session_url: 'http://127.0.0.1:9000/logout', allow_http_localhost: true
 	}}, {}));
+	const localhostRedirect = new URL(OIDC.buildOidcLogoutLocation({ client_id: 'client', logout: {
+		enabled: true,
+		end_session_url: 'https://idp.example/logout',
+		post_logout_redirect_uri: 'http://127.0.0.1:3012',
+		allow_http_localhost: true
+	}}, {}));
+	assert.equal(localhostRedirect.searchParams.get('post_logout_redirect_uri'), 'http://127.0.0.1:3012');
+	const literalRedirect = new URL(OIDC.buildOidcLogoutLocation({ client_id: 'client', logout: {
+		enabled: true,
+		end_session_url: 'https://idp.example/logout',
+		post_logout_redirect_uri: 'https://cron.example'
+	}}, {}));
+	assert.equal(literalRedirect.searchParams.get('post_logout_redirect_uri'), 'https://cron.example');
+	assert.throws(() => OIDC.buildOidcLogoutLocation({ logout: {
+		enabled: true,
+		end_session_url: 'https://idp.example/logout',
+		post_logout_redirect_uri: ' https://cron.example'
+	}}, {}), /whitespace/);
+});
+
+test('OIDC provider allows localhost HTTP without disabling TLS verification', () => {
+	const provider = OIDC.validateOidcProviderConfig({
+		client_id: 'cronicle-edge',
+		issuer: 'http://localhost:9000/application/o/cronicle/',
+		jwks_url: 'http://127.0.0.1:9000/application/o/cronicle/jwks/',
+		insecure: false,
+		allow_http_localhost: true
+	});
+	assert.equal(provider.issuer, 'http://localhost:9000/application/o/cronicle/');
+	assert.throws(() => OIDC.validateOidcProviderConfig({
+		client_id: 'cronicle-edge',
+		issuer: 'http://idp.example/application/o/cronicle/',
+		jwks_url: 'http://idp.example/application/o/cronicle/jwks/',
+		insecure: false,
+		allow_http_localhost: true
+	}), /HTTPS/);
 });
 
 test('OIDC session metadata is minimal and validates subject consistency', () => {
@@ -226,19 +283,19 @@ test('OIDC session metadata is minimal and validates subject consistency', () =>
 		client_id: 'cronicle-edge', issuer: 'https://idp.example/realm', jwks_url: 'https://idp.example/jwks',
 		logout: { enabled: true }
 	};
-	const metadata = OIDC.buildSessionMetadata(oauth, { sub: 'alice-id', sid: 'sid-1' }, { sub: 'alice-id' },
+	const metadata = OIDC.buildOidcSessionMetadata(oauth, { sub: 'alice-id', sid: 'sid-1' }, { sub: 'alice-id' },
 		'id.token', 'session-secret');
 	assert.equal(metadata.auth_provider, 'oidc');
 	assert.equal(metadata.oidc_issuer, 'https://idp.example/realm');
 	assert.equal(metadata.oidc_subject, 'alice-id');
 	assert.equal(metadata.oidc_client_id, 'cronicle-edge');
 	assert.equal(metadata.oidc_sid, 'sid-1');
-	assert.equal(OIDC.decryptSessionToken(metadata.oidc_id_token_hint_enc, 'session-secret'), 'id.token');
+	assert.equal(OIDC.decryptOidcSessionToken(metadata.oidc_id_token_hint_enc, 'session-secret'), 'id.token');
 	assert.equal(JSON.stringify(metadata).includes('id.token'), false);
 	assert.equal(metadata.access_token, undefined);
 	assert.equal(metadata.refresh_token, undefined);
-	assert.throws(() => OIDC.buildSessionMetadata(oauth, { sub: 'one' }, { sub: 'two' }, 'id.token', 'session-secret'), /does not match/);
-	assert.throws(() => OIDC.buildSessionMetadata(oauth, { sub: 'alice-id' }, { sub: 'alice-id' }, null,
+	assert.throws(() => OIDC.buildOidcSessionMetadata(oauth, { sub: 'one' }, { sub: 'two' }, 'id.token', 'session-secret'), /does not match/);
+	assert.throws(() => OIDC.buildOidcSessionMetadata(oauth, { sub: 'alice-id' }, { sub: 'alice-id' }, null,
 		'session-secret'), /requires a verified ID Token/);
 });
 
@@ -270,7 +327,7 @@ before(async () => {
 		client_id: 'cronicle-edge',
 		issuer: `http://127.0.0.1:${address.port}/issuer`,
 		jwks_url: `http://127.0.0.1:${address.port}/jwks`,
-		insecure: true,
+		insecure: false,
 		allow_http_localhost: true,
 		allowed_algs: ['RS256'],
 		jwks_cooldown_seconds: 0,
@@ -291,7 +348,7 @@ async function logoutToken(overrides, signingKey, kid) {
 		iat: now,
 		exp: now + 300,
 		jti: 'jti-' + Math.random(),
-		events: { [OIDC.BACKCHANNEL_EVENT]: {} },
+		events: { [OIDC.OIDC_BACKCHANNEL_EVENT]: {} },
 		sid: 'sid-1'
 	}, overrides || {});
 	return new jose.SignJWT(values)
@@ -300,39 +357,53 @@ async function logoutToken(overrides, signingKey, kid) {
 }
 
 test('Verified ID Tokens require iat, audience, subject, and the login nonce', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const now = Math.floor(Date.now() / 1000);
 	const token = await new jose.SignJWT({
 		iss: oauth.issuer, aud: oauth.client_id, sub: 'subject-id', iat: now, exp: now + 300, nonce: 'login-nonce'
 	}).setProtectedHeader({ alg: 'RS256', kid: 'key-1' }).sign(key1.privateKey);
-	await assert.doesNotReject(OIDC.verifyIdToken(token, oauth, 'login-nonce'));
-	await assert.rejects(OIDC.verifyIdToken(token, oauth, 'wrong-nonce'), /nonce/);
+	await assert.doesNotReject(OIDC.verifyOidcIdToken(token, oauth, 'login-nonce'));
+	await assert.rejects(OIDC.verifyOidcIdToken(token, oauth, 'wrong-nonce'), /nonce/);
 });
 
 test('Back-channel JWT validation accepts a valid token and rejects invalid claims', async () => {
-	OIDC._resetCachesForTests();
-	await assert.doesNotReject(OIDC.verifyLogoutToken(await logoutToken(), oauth));
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ iss: 'https://wrong.example' }), oauth));
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ aud: 'wrong-client' }), oauth));
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ events: {} }), oauth), /events/);
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ nonce: 'forbidden' }), oauth), /nonce/);
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ iat: Math.floor(Date.now() / 1000) - 600 }), oauth), /iat/);
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ exp: Math.floor(Date.now() / 1000) - 10 }), oauth));
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({ sid: undefined, sub: undefined }), oauth), /sid or sub/);
+	OIDC.resetOidcCachesForTests();
+	await assert.doesNotReject(OIDC.verifyOidcLogoutToken(await logoutToken(), oauth));
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ iss: 'https://wrong.example' }), oauth));
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ aud: 'wrong-client' }), oauth));
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ events: {} }), oauth), /events/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ nonce: 'forbidden' }), oauth), /nonce/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ iat: Math.floor(Date.now() / 1000) - 600 }), oauth), /iat/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ exp: Math.floor(Date.now() / 1000) - 10 }), oauth));
+	await assert.doesNotReject(OIDC.verifyOidcLogoutToken(await logoutToken({ exp: undefined }), oauth));
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ sid: undefined, sub: undefined }), oauth), /sid or sub/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ sid: 123, sub: 'subject-id' }), oauth), /sid/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ sid: 'sid-1', sub: 123 }), oauth), /subject/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ sid: undefined, sub: 'ę' }), oauth), /subject/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ aud: [oauth.client_id, 'other'] }), oauth), /audience/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({
+		aud: [oauth.client_id, 'other'], azp: 'other'
+	}), oauth), /audience/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({
+		aud: [oauth.client_id, 'other'], azp: oauth.client_id
+	}), oauth), /audience/);
+	await assert.doesNotReject(OIDC.verifyOidcLogoutToken(await logoutToken({ aud: [oauth.client_id] }), oauth));
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ sid: 'ę' }), oauth), /sid/);
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({ sid: 'line\nbreak' }), oauth), /sid/);
 });
 
 test('Back-channel JWT validation rejects bad signatures and unknown keys', async () => {
-	OIDC._resetCachesForTests();
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({}, key2, 'key-1'), oauth));
-	await assert.rejects(OIDC.verifyLogoutToken(await logoutToken({}, key2, 'unknown-key'), oauth));
+	OIDC.resetOidcCachesForTests();
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({}, key2, 'key-1'), oauth));
+	await assert.rejects(OIDC.verifyOidcLogoutToken(await logoutToken({}, key2, 'unknown-key'), oauth));
 });
 
 test('Remote JWKS cache supports signing-key rotation', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	jwks = [jwk1];
-	await OIDC.verifyLogoutToken(await logoutToken({ jti: 'rotation-1' }), oauth);
+	await OIDC.verifyOidcLogoutToken(await logoutToken({ jti: 'rotation-1' }), oauth);
 	jwks = [jwk2];
-	await OIDC.verifyLogoutToken(await logoutToken({ jti: 'rotation-2' }, key2, 'key-2'), oauth);
+	await OIDC.verifyOidcLogoutToken(await logoutToken({ jti: 'rotation-2' }, key2, 'key-2'), oauth);
 	jwks = [jwk1];
 });
 
@@ -340,8 +411,8 @@ test('Session index covers sid and sub and is cleaned on local deletion', async 
 	const { user, storage } = makeUser(oauth);
 	const session = makeSession('session-1', oauth.issuer, 'subject-1', 'sid-1');
 	await call(user, 'storeOidcSession', session);
-	const sidKey = OIDC.indexKey(oauth.issuer, 'sid', 'sid-1');
-	const subKey = OIDC.indexKey(oauth.issuer, 'sub', 'subject-1');
+	const sidKey = OIDC.getOidcIndexKey(oauth.issuer, 'sid', 'sid-1');
+	const subKey = OIDC.getOidcIndexKey(oauth.issuer, 'sub', 'subject-1');
 	assert.equal(storage.records.get(sidKey).sessions['session-1'], session.expires);
 	assert.equal(storage.records.get(subKey).sessions['session-1'], session.expires);
 	await call(user, 'deleteSessionRecord', session);
@@ -355,23 +426,24 @@ test('Expired OIDC sessions clean their index records', async () => {
 	const session = makeSession('session-expired', oauth.issuer, 'subject-expired', 'sid-expired');
 	await call(user, 'storeOidcSession', session);
 	await storage.expireNow('sessions/' + session.id);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
 });
 
 test('Back-channel sid logout deletes only the matching session', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage } = makeUser(oauth);
 	await call(user, 'storeOidcSession', makeSession('sid-match', oauth.issuer, 'subject-1', 'sid-1'));
 	await call(user, 'storeOidcSession', makeSession('sid-other', oauth.issuer, 'subject-1', 'sid-2'));
 	const response = await invokeBackchannel(user, await logoutToken({ sid: 'sid-1', sub: 'subject-1' }));
 	assert.equal(response.status, 200);
+	assert.equal(response.headers['Cache-Control'], 'no-store');
 	assert.equal(recordExists(storage, 'sessions/sid-match'), false);
 	assert.equal(recordExists(storage, 'sessions/sid-other'), true);
 });
 
 test('Back-channel sub logout deletes all user sessions', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage } = makeUser(oauth);
 	await call(user, 'storeOidcSession', makeSession('sub-one', oauth.issuer, 'subject-all', 'sid-a'));
 	await call(user, 'storeOidcSession', makeSession('sub-two', oauth.issuer, 'subject-all', 'sid-b'));
@@ -383,7 +455,7 @@ test('Back-channel sub logout deletes all user sessions', async () => {
 });
 
 test('Back-channel replay and unknown target are idempotent HTTP 200', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage } = makeUser(oauth);
 	const session = makeSession('replay-session', oauth.issuer, 'subject-replay', 'sid-replay');
 	await call(user, 'storeOidcSession', session);
@@ -394,26 +466,42 @@ test('Back-channel replay and unknown target are idempotent HTTP 200', async () 
 	assert.equal((await invokeBackchannel(user, await logoutToken({ sid: 'unknown-sid', sub: undefined }))).status, 200);
 });
 
+test('Back-channel logout without exp uses a finite replay and revocation window', async () => {
+	OIDC.resetOidcCachesForTests();
+	const { user, storage } = makeUser(oauth);
+	const now = Math.floor(Date.now() / 1000);
+	const token = await logoutToken({ exp: undefined, sid: 'sid-no-exp', sub: undefined, jti: 'jti-no-exp' });
+	assert.equal((await invokeBackchannel(user, token)).status, 200);
+
+	const replayKey = OIDC.getOidcReplayKey(oauth.issuer, 'jti-no-exp');
+	const markerKey = OIDC.getOidcRevocationKey(oauth.issuer, oauth.client_id, 'sid', 'sid-no-exp');
+	[replayKey, markerKey].forEach(function(key) {
+		const expires = storage.expirations.get(key);
+		assert.equal(Number.isFinite(expires), true);
+		assert.ok(expires > now && expires <= now + 310);
+	});
+});
+
 test('Back-channel logout blocks a delayed callback from recreating the revoked session', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage } = makeUser(oauth);
 	const token = await logoutToken({ sid: 'sid-delayed', sub: 'subject-delayed', jti: 'delayed-callback-jti' });
 	const response = await invokeBackchannel(user, token);
 	assert.equal(response.status, 200);
 
-	const markerKey = OIDC.revocationKey(oauth.issuer, oauth.client_id, 'sid', 'sid-delayed');
+	const markerKey = OIDC.getOidcRevocationKey(oauth.issuer, oauth.client_id, 'sid', 'sid-delayed');
 	assert.equal(recordExists(storage, markerKey), true);
 	assert.ok(storage.expirations.get(markerKey) > Math.floor(Date.now() / 1000));
 
 	const session = makeSession('delayed-session', oauth.issuer, 'subject-delayed', 'sid-delayed');
 	await assert.rejects(call(user, 'storeOidcSession', session), /revoked before it could be stored/);
 	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
 });
 
 test('Concurrent callback and back-channel logout cannot leave an active session', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage } = makeUser(oauth);
 	const session = makeSession('callback-race-session', oauth.issuer, 'subject-race', 'sid-race');
 	const token = await logoutToken({ sid: 'sid-race', sub: 'subject-race', jti: 'callback-race-jti' });
@@ -424,16 +512,52 @@ test('Concurrent callback and back-channel logout cannot leave an active session
 	assert.equal(results[1].status, 'fulfilled');
 	assert.equal(results[1].value.status, 200);
 	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
 	assert.equal(recordExists(storage,
-		OIDC.revocationKey(oauth.issuer, oauth.client_id, 'sid', session.oidc_sid)), true);
+		OIDC.getOidcRevocationKey(oauth.issuer, oauth.client_id, 'sid', session.oidc_sid)), true);
+});
+
+test('Back-channel logout cannot be undone by an in-flight session resume', async () => {
+	OIDC.resetOidcCachesForTests();
+	const { user, storage } = makeUser(oauth);
+	const session = makeSession('resume-race-session', oauth.issuer, 'subject-resume', 'sid-resume');
+	await call(user, 'storeOidcSession', session);
+
+	user.config = { get: () => 30 };
+	storage.config = { get: () => false };
+	user.getClientInfo = () => ({});
+	user.doError = (code, description, callback) => callback({ code: 1, description });
+	user.loadSession = (args, callback) => callback(null, storage.clone(session), {
+		username: 'alice', email: 'alice@example.test', full_name: 'Alice', active: 1
+	});
+
+	const token = await logoutToken({ sid: 'sid-resume', sub: 'subject-resume', jti: 'resume-race-jti' });
+	let logoutResponse;
+	user.fireHook = function(name, args, callback) {
+		if (name !== 'before_resume_session') return callback && callback();
+		invokeBackchannel(user, token).then(function(response) {
+			logoutResponse = response;
+			callback();
+		}, callback);
+	};
+
+	const response = await new Promise((resolve) => user.api_resume_session({
+		request: { headers: { 'user-agent': 'test' } }, params: {}, query: {}, cookies: {}, ip: '127.0.0.1',
+		_oidc_secrets: { session_id: session.id }
+	}, resolve));
+
+	assert.equal(logoutResponse.status, 200);
+	assert.equal(response.code, 1);
+	assert.match(response.description, /revoked before it could be stored/);
+	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
 });
 
 test('Expired logout revocation markers do not block a later valid session', async () => {
 	const { user, storage } = makeUser(oauth);
 	const session = makeSession('post-expiry-session', oauth.issuer, 'subject-expiry', 'sid-expiry');
-	const markerKey = OIDC.revocationKey(oauth.issuer, oauth.client_id, 'sid', session.oidc_sid);
+	const markerKey = OIDC.getOidcRevocationKey(oauth.issuer, oauth.client_id, 'sid', session.oidc_sid);
 	await new Promise((resolve, reject) => storage.put(markerKey, {
 		created: Math.floor(Date.now() / 1000) - 600,
 		expires: Math.floor(Date.now() / 1000) - 1
@@ -443,25 +567,28 @@ test('Expired logout revocation markers do not block a later valid session', asy
 	assert.equal(recordExists(storage, 'sessions/' + session.id), true);
 });
 
-test('Back-channel endpoint returns 400 for invalid input and 405 for non-POST', async () => {
-	OIDC._resetCachesForTests();
+test('Back-channel endpoint returns spec-compliant HTTP 400 errors', async () => {
+	OIDC.resetOidcCachesForTests();
 	const { user } = makeUser(oauth);
-	assert.equal((await invokeBackchannel(user, '')).status, 400);
+	const invalid = await invokeBackchannel(user, '');
+	assert.equal(invalid.status, 400);
+	assert.equal(invalid.body.error, 'invalid_request');
+	assert.equal(invalid.headers['Cache-Control'], 'no-store');
 	assert.equal((await invokeBackchannel(user, await logoutToken(), { contentType: 'application/json' })).status, 400);
-	assert.equal((await invokeBackchannel(user, await logoutToken(), { method: 'GET' })).status, 405);
+	assert.equal((await invokeBackchannel(user, await logoutToken(), { method: 'GET' })).status, 400);
 	assert.equal((await invokeBackchannel(user, 'x'.repeat(17000))).status, 400);
 	assert.equal((await invokeBackchannel(user, await logoutToken(), { contentLength: 20000 })).status, 400);
 });
 
 test('Back-channel endpoint surfaces configuration, JWKS, and storage failures', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const misconfigured = Object.assign({}, oauth, { jwks_url: undefined });
-	assert.equal((await invokeBackchannel(makeUser(misconfigured).user, await logoutToken())).status, 500);
+	assert.equal((await invokeBackchannel(makeUser(misconfigured).user, await logoutToken())).status, 400);
 
 	const unavailable = Object.assign({}, oauth, { jwks_url: 'http://127.0.0.1:1/jwks' });
-	assert.equal((await invokeBackchannel(makeUser(unavailable).user, await logoutToken())).status, 503);
+	assert.equal((await invokeBackchannel(makeUser(unavailable).user, await logoutToken())).status, 400);
 
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage, logs } = makeUser(oauth);
 	const originalPut = storage.put.bind(storage);
 	storage.put = function(key, value, callback) {
@@ -469,12 +596,12 @@ test('Back-channel endpoint surfaces configuration, JWKS, and storage failures',
 		return originalPut(key, value, callback);
 	};
 	const token = await logoutToken({ sid: 'unknown-storage-target', jti: 'storage-failure-jti' });
-	assert.equal((await invokeBackchannel(user, token)).status, 500);
+	assert.equal((await invokeBackchannel(user, token)).status, 400);
 	assert.equal(logs.join('\n').includes(token), false);
 });
 
 test('Concurrent local and back-channel logout leaves no session or index', async () => {
-	OIDC._resetCachesForTests();
+	OIDC.resetOidcCachesForTests();
 	const { user, storage } = makeUser(oauth);
 	const session = makeSession('concurrent-session', oauth.issuer, 'subject-concurrent', 'sid-concurrent');
 	await call(user, 'storeOidcSession', session);
@@ -484,8 +611,8 @@ test('Concurrent local and back-channel logout leaves no session or index', asyn
 		invokeBackchannel(user, token)
 	]);
 	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
-	assert.equal(recordExists(storage, OIDC.indexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sid', session.oidc_sid)), false);
+	assert.equal(recordExists(storage, OIDC.getOidcIndexKey(oauth.issuer, 'sub', session.oidc_subject)), false);
 });
 
 test('Old sessions remain locally deletable and are ignored by OIDC indexes', async () => {
@@ -545,6 +672,25 @@ test('Password sessions never trigger the configured OIDC logout', async () => {
 	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
 });
 
+test('RP logout does not send an ID Token to a different configured provider', async () => {
+	const rpOauth = Object.assign({}, oauth, {
+		logout: { enabled: true, end_session_url: 'https://idp.example/logout' }
+	});
+	const { user, storage } = makeUser(rpOauth);
+	const session = Object.assign(makeSession('old-provider-session', 'https://old-idp.example/', 'subject-old', 'sid-old'), {
+		oidc_id_token_hint_enc: OIDC.encryptOidcSessionToken('old.id.token', 'test-ticket-secret')
+	});
+	await call(user, 'storeOidcSession', session);
+	await new Promise((resolve, reject) => storage.put('users/alice', {
+		username: 'alice', active: 1, email: 'a@example.com', full_name: 'Alice'
+	}, (err) => err ? reject(err) : resolve()));
+	const response = await invokeLocalLogout(user, session.id);
+	assert.equal(response.code, 0);
+	assert.equal(response.logout_location, undefined);
+	assert.match(response.logout_warning, /Local logout succeeded/);
+	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
+});
+
 test('RP logout deletes local state before returning an unreachable IdP URL', async () => {
 	const rpOauth = Object.assign({}, oauth, {
 		logout: {
@@ -555,7 +701,7 @@ test('RP logout deletes local state before returning an unreachable IdP URL', as
 	});
 	const { user, storage, logs } = makeUser(rpOauth);
 	const session = Object.assign(makeSession('rp-session', oauth.issuer, 'subject-rp', 'sid-rp'), {
-		oidc_id_token_hint_enc: OIDC.encryptSessionToken('secret.id.token', 'test-ticket-secret')
+		oidc_id_token_hint_enc: OIDC.encryptOidcSessionToken('secret.id.token', 'test-ticket-secret')
 	});
 	await call(user, 'storeOidcSession', session);
 	assert.equal(JSON.stringify(storage.records.get('sessions/' + session.id)).includes('secret.id.token'), false);
@@ -577,7 +723,7 @@ test('RP logout deletes local state before returning an unreachable IdP URL', as
 	assert.match(response.logout_location, /^\/api\/user\/oidc_logout_redirect\?ticket=/);
 	assert.equal(response.logout_location.includes('secret.id.token'), false);
 	const ticket = new URL(response.logout_location, 'https://cron.example').searchParams.get('ticket');
-	const destination = OIDC.consumeLogoutTicket(ticket, 'test-ticket-secret').location;
+	const destination = OIDC.consumeOidcLogoutTicket(ticket, 'test-ticket-secret').location;
 	assert.match(destination, /^https:\/\/127\.0\.0\.1:9\/logout/);
 	assert.equal(destination.includes('attacker.example'), false);
 	assert.equal(recordExists(storage, 'sessions/' + session.id), false);
@@ -588,7 +734,7 @@ test('RP logout deletes local state before returning an unreachable IdP URL', as
 test('Backend logout_location consumes its ticket once and redirects to the configured IdP', async () => {
 	const { user } = makeUser(oauth);
 	const destination = 'https://idp.example/logout?id_token_hint=secret.id.token';
-	const ticket = OIDC.createLogoutTicket(destination, 'test-ticket-secret', 60);
+	const ticket = OIDC.createOidcLogoutTicket(destination, 'test-ticket-secret', 60);
 	let redirect;
 	const args = {
 		request: { method: 'GET', url: '/api/user/oidc_logout_redirect?ticket=' + ticket, headers: {} },
